move fetchImageData function to utils
because, it will also be needed on the options page in addition to the
extension popup.
