+++ src/popup/infoPopupModule.js
@@ -88,15 +88,6 @@ Creator Link: ${creatorUrl}\n\n
 ${HtmlAttribution}`;
 }
 
-export async function fetchImageData(imageId) {
-  const url = `https://api.creativecommons.engineering/v1/images/${imageId}`;
-  const data = await fetch(url);
-  const responseCode = data.status;
-  const res = await data.json();
-
-  return [res, responseCode];
-}
-
 function downloadImage(imageUrl, imageName) {
   const x = new XMLHttpRequest();
   x.open('GET', imageUrl, true);
