+++ src/utils.js
@@ -87,3 +87,12 @@ export function removeChildNodes(targetNode) {
     targetNode.removeChild(targetNode.lastChild);
   }
 }
+
+export async function fetchImageData(imageId) {
+  const url = `https://api.creativecommons.engineering/v1/images/${imageId}`;
+  const data = await fetch(url);
+  const responseCode = data.status;
+  const res = await data.json();
+
+  return [res, responseCode];
+}
